requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
deta==1.1.0
idna==3.4
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.21.0
requests==2.28.1
update-checker==0.18.0
urllib3==1.26.12
websocket-client==1.4.1
